Move Ensure/FreeSignalAlternateStack to CPalThread class.
Frees the alternate stack a little sooner during thread termination.
